agent/pyproject.toml
[project]
name = "bancobot-agent"
version = "0.1.0"
description = "BancoBot AI — Agent (Agno + FastAPI + Redis)"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "openai>=1.56.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
